requirements.txt
dash==2.3.0
mysql_connector_repackaged==0.3.1
neo4j==4.4.2
numpy==1.22.2
pandas==1.4.1
plotly==5.6.0
pymongo==4.0.1
Werkzeug==2.1.0
